requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
ca-certificates=2020.4.5.1=hecc5488_0
certifi=2020.4.5.1=py37hc8dfbb8_0
cffi=1.14.0=py37h356ff06_0
chardet=3.0.4=py37hc8dfbb8_1006
cryptography=2.8=py37he655712_2
cycler=0.10.0=py_2
freetype=2.10.1=h8da9a1a_0
idna=2.9=py_1
joblib=0.14.1=py_0
kiwisolver=1.2.0=py37ha1cc60f_0
libblas=3.8.0=16_openblas
libcblas=3.8.0=16_openblas
libcxx=10.0.0=0
libffi=3.2.1=h4a8c4bd_1007
libgfortran=4.0.0=2
liblapack=3.8.0=16_openblas
libopenblas=0.3.9=h3d69b6c_0
libpng=1.6.37=hbbe82c9_1
llvm-openmp=9.0.1=h28b9765_2
matplotlib-base=3.2.1=py37hddda452_0
ncurses=6.1=h0a44026_1002
nose=1.3.7=py37hc8dfbb8_1004
numpy=1.18.1=py37h7687784_1
openssl=1.1.1f=h0b31af3_0
pandas=1.0.3=py37h94625e5_0
pip=20.0.2=py_2
pycparser=2.20=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py37hc8dfbb8_1
python=3.7.6=h90870a6_5_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytz=2019.3=py_0
readline=8.0=hcfe32e1_0
requests=2.23.0=pyh8c360ce_2
scikit-learn=0.22.2.post1=py37h3dc85bc_0
scipy=1.4.1=py37hce1b9e5_2
setuptools=46.1.3=py37hc8dfbb8_0
six=1.14.0=py_1
sqlite=3.30.1=h93121df_0
tk=8.6.10=hbbe82c9_0
tornado=6.0.4=py37h9bfed18_1
urllib3=1.25.7=py37hc8dfbb8_1
wfdb=2.2.1=py_1
wheel=0.34.2=py_1
xz=5.2.5=h0b31af3_0
zlib=1.2.11=h0b31af3_1006
